Changing how table names are modified if camelCase
Not all tables are snake_case. If the source table is CamelCase, then use that as the ModuleName

If the table has a capital letter in it, use the current format as it is most likely camel case. If it does not have a capital letter, use the existing method of converting a snake_case_name into CamelCase

Signed-off-by: Chris Desch <cdesch@gmail.com>